+++ Assets/Shaders/PostProcessInclude.hlsl
@@ -8,13 +8,13 @@ half4 _MainTex_TexelSize;
 TEXTURE2D( _CameraDepthTexture); SAMPLER(sampler_CameraDepthTexture);
 half4 _CameraDepthTexture_TexelSize;
 
-float4 SampleMainTex(half2 uv){return  SAMPLE_TEXTURE2D(_MainTex,sampler_MainTex,uv);}
-float SampleRawDepth(half2 uv){return SAMPLE_DEPTH_TEXTURE(_CameraDepthTexture,sampler_CameraDepthTexture,uv).r;}
-float SampleEyeDepth(half2 uv){return RawToEyeDepth(SampleRawDepth(uv));}
-float Sample01Depth(half2 uv){return RawTo01Depth(SampleRawDepth(uv));}
-float3 TransformNDCToWorld(half2 uv){return TransformNDCToWorld(uv,SampleRawDepth(uv));}
+float4 SampleMainTex(float2 uv){return  SAMPLE_TEXTURE2D(_MainTex,sampler_MainTex,uv);}
+float SampleRawDepth(float2 uv){return SAMPLE_DEPTH_TEXTURE(_CameraDepthTexture,sampler_CameraDepthTexture,uv).r;}
+float SampleEyeDepth(float2 uv){return RawToEyeDepth(SampleRawDepth(uv));}
+float Sample01Depth(float2 uv){return RawTo01Depth(SampleRawDepth(uv));}
+float3 TransformNDCToWorld(float2 uv){return TransformNDCToWorld(uv,SampleRawDepth(uv));}
 
-float3 WorldSpaceNormalFromDepth(half2 uv,inout float3 positionWS,inout half depth)
+float3 WorldSpaceNormalFromDepth(float2 uv,inout float3 positionWS,inout half depth)
 {
     depth=SampleRawDepth(uv);
     positionWS=TransformNDCToWorld(uv,depth);
@@ -22,7 +22,7 @@ float3 WorldSpaceNormalFromDepth(half2 uv,inout float3 positionWS,inout half dep
     float3 position2=TransformNDCToWorld(uv+_MainTex_TexelUp);
     return normalize(cross(position2-positionWS,position1-positionWS));
 }
-half3 ClipSpaceNormalFromDepth(half2 uv)
+half3 ClipSpaceNormalFromDepth(float2 uv)
 {
     half depth = SampleEyeDepth(uv);
     half depth1 = SampleEyeDepth(uv + _MainTex_TexelRight);
@@ -38,13 +38,13 @@ half luminance(half3 color){ return 0.299h * color.r + 0.587h * color.g + 0.114h
 struct a2v_img
 {
     half3 positionOS : POSITION;
-    half2 uv : TEXCOORD0;
+    float2 uv : TEXCOORD0;
 };
 
 struct v2f_img
 {
     float4 positionCS : SV_Position;
-    half2 uv : TEXCOORD0;
+    float2 uv : TEXCOORD0;
 };
 
 v2f_img vert_img(a2v_img v)
